Create arduino led and sensor project
This project is about LED-Distance Indicator by using an ultrasonic sensor and Arduino system software. Arduino is basically open-source electronics prototyping platform based on a microcontroller board, and development environment for writing software for the board. So here we connected our ultrasonic sensor with Arduino and, we designed an Arduino code so that it works accordingly. If we move an object towards sensor then one by one LED lights will get turned on and if we move the object away from sensor then one by one LED lights will get turned off.